fix: The edit board form now works correctly
Previously the form didn't perform the update on a Board since the utility function to check if a Board is valid was checking for tasks property, when in reality we don't send the tasks in the edit board form, only the name and description

The 'darker-blue' color has been added to the tailwind.config.ts file to provide a darker shade of blue for the Edit Board button and allow it to change when the button is hovered. Additionally, the error message in the board route has been updated to reflect that the error occurred while updating the board, not the task.